backend/requirements.txt
# Z-Image AI Service — Python 3.11+
fastapi>=0.109.0,<0.115.0
uvicorn[standard]>=0.27.0,<0.32.0
pydantic>=2.6.0,<2.10.0
pydantic-settings>=2.2.0,<2.6.0
python-multipart>=0.0.9

# ML
torch>=2.2.0,<2.6.0
diffusers>=0.36.0,<0.41.0
transformers>=4.36.0
accelerate>=0.26.0
safetensors>=0.4.0
Pillow>=10.0.0,<11.0.0
# 한국어 LoRA 파인튜닝 (서버 기동 시 train_lora.py 서브프로세스에서 사용)
datasets>=2.16.0
peft>=0.10.0
trl>=0.8.0
